+++ plugins/postcss-custom-media/CHANGELOG.md
@@ -1,6 +1,8 @@
-# 4.0.0 - Unrelease
+# 4.0.0 - 2015-05-17
 
-- Changed: warning messages are not sent via postcss API (^4.1.0)
+- Changed: warning messages are now sent via postcss messages api (^4.1.0)
+- Added: automatic custom media `--` prefixing
+([#11](https://github.com/postcss/postcss-custom-media/issues/11))
 - Added: `preserve` allows you to preserve custom media query defintions
 - Added: `appendExtensions` allows you (when `preserve` is truthy) to append your extensions as media queries